计算业务可用度

+++ src/main/java/com/gy/businessCore/entity/BusinessEntity.java
@@ -24,13 +24,13 @@ public class BusinessEntity {
     private String name;
 
     @Column(name = "busy_score")
-    private float busy_score;
+    private double busy_score;
 
     @Column(name = "health_score")
-    private float health_score;
+    private double health_score;
 
     @Column(name = "available_score")
-    private float available_score;
+    private double available_score;
 
     @Column(name = "cluster")
     private String cluster;
